backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
PyJWT
drf-yasg
djangorestframework-simplejwt
django-filter
